requirements.txt
Flask
opencv-python
numpy
werkzeug
Pillow